backend/requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
atomicwrites==1.4.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
contourpy==1.0.5
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cycler==0.11.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.1.2
django-ckeditor==6.5.1
django-constance==2.9.1
django-cors-headers==3.13.0
django-filter==22.1
django-js-asset==2.0.0
django-picklefield==3.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-versatileimagefield==2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
fonttools==4.37.4
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.4
lineage==4.3.0
django-lineage==0.2.0
MarkupSafe==2.1.1
matplotlib==3.6.1
numpy==1.23.4
oauthlib==3.2.1
openapi-codec==1.3.2
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
psycopg2==2.9.4
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.4
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
snps==2.6.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.5
uritemplate==4.1.1
urllib3==1.26.12
wrapt==1.14.1
gunicorn==20.1.0
drf-registration==0.1.3
